refactor: 공통 메서드 상위 클래스로 이동
