+++ src/main/java/chess/domain/chessboard/state/piece/Bishop.java
@@ -1,9 +1,7 @@
 package chess.domain.chessboard.state.piece;
 
 import chess.domain.chessboard.Coordinate;
-import chess.domain.chessboard.Square;
 import chess.domain.chessboard.state.Team;
-import java.util.ArrayList;
 import java.util.List;
 
 public final class Bishop extends Piece {
@@ -22,63 +20,8 @@ public final class Bishop extends Piece {
         return negativeDiagonalRoute(from, to);
     }
 
-    private List<Coordinate> negativeDiagonalRoute(final Coordinate from, final Coordinate to) {
-        final int distance = from.calculateRankDistance(to);
-
-        if (distance < 0) {
-            return negativeDiagonalRouteBySign(from, distance, -1);
-        }
-
-        return negativeDiagonalRouteBySign(from, distance, 1);
-    }
-
-    private List<Coordinate> negativeDiagonalRouteBySign(final Coordinate from, final int distance, final int sign) {
-        final List<Coordinate> route = new ArrayList<>();
-        for (int i = 1; i <= sign * distance; i++) {
-            route.add(from.negativeDiagonalMove(sign * i));
-        }
-        return route;
-    }
-
-    private List<Coordinate> positiveDiagonalRoute(final Coordinate from, final Coordinate to) {
-        final int distance = from.calculateFileDistance(to);
-
-        if (distance < 0) {
-            return positiveDiagonalRouteBySign(from, distance, -1);
-        }
-
-        return positiveDiagonalRouteBySign(from, distance, 1);
-    }
-
-    private List<Coordinate> positiveDiagonalRouteBySign(final Coordinate from, final int distance, final int sign) {
-        final List<Coordinate> route = new ArrayList<>();
-        for (int i = 1; i <= sign * distance; i++) {
-            route.add(from.positiveDiagonalMove(sign * i));
-        }
-        return route;
-    }
-
     private void validatePossibleDestination(final Coordinate from, final Coordinate to) {
         if (!(from.isPositiveDiagonal(to) || from.isNegativeDiagonal(to))) {
-            throwCanNotMoveException();
-        }
-    }
-
-    @Override
-    public void canMove(final List<Square> routeSquares) {
-        final int lastIndex = routeSquares.size() - 1;
-        final Square lastSquare = routeSquares.get(lastIndex);
-        if (lastSquare.isSameTeam(this)) {
-            throwCanNotMoveException();
-        }
-
-        for (Square curSquare : routeSquares.subList(0, lastIndex)) {
-            checkSquareEmpty(curSquare);
-        }
-    }
-
-    private void checkSquareEmpty(final Square curSquare) {
-        if (!curSquare.isEmpty()) {
             throwCanNotMoveException();
         }
     }
